add Filter stateless component
